change: files location, refactor: TextField
